litellm-lusofona/contact/src/requirements.txt
# Minimal dependencies for the simplified bulk email sender
google-api-python-client>=2.0.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=0.5.0
python-dotenv>=0.19.0
